+++ core/lib/Thelia/Command/ModuleDeactivateCommand.php
@@ -61,8 +61,7 @@ class ModuleDeactivateCommand extends BaseModuleGenerate
 
         try {
             $event = new ModuleToggleActivationEvent($module->getId());
-            $dispatcher = $this->getContainer()->get('event_dispatcher');
-            $dispatcher->dispatch(TheliaEvents::MODULE_TOGGLE_ACTIVATION, $event);
+            $this->getDispatcher()->dispatch(TheliaEvents::MODULE_TOGGLE_ACTIVATION, $event);
         } catch (\Exception $e) {
             throw new \RuntimeException(sprintf("Deactivation fail with Exception : [%d] %s", $e->getCode(), $e->getMessage()));
         }