+++ core/lib/Thelia/Core/Thelia.php
@@ -42,6 +42,7 @@ use Thelia\Core\Template\ParserInterface;
 use Thelia\Core\Template\TemplateDefinition;
 use Thelia\Core\Template\TemplateHelper;
 use Thelia\Log\Tlog;
+use Thelia\Model\Map\ProductTableMap;
 use Thelia\Model\Module;
 use Thelia\Model\ModuleQuery;
 
@@ -75,16 +76,26 @@ class Thelia extends Kernel
             new DatabaseConfiguration(),
             Yaml::parse(THELIA_CONF_DIR . 'database.yml')
         );
+
+        /** @var \Propel\Runtime\ServiceContainer\StandardServiceContainer $serviceContainer */
         $serviceContainer = Propel::getServiceContainer();
-        $serviceContainer->setAdapterClass('thelia', 'mysql');
         $serviceContainer->setDefaultDatasource('thelia');
 
         $manager = new ConnectionManagerSingle();
         $manager->setConfiguration($definePropel->getConfig());
+        $manager->setName('thelia');
+
         $serviceContainer->setConnectionManager('thelia', $manager);
-        $con = Propel::getConnection(\Thelia\Model\Map\ProductTableMap::DATABASE_NAME);
+        $serviceContainer->setAdapterClass('thelia', 'mysql');
+
+        /** @var ConnectionWrapper $con */
+        $con = Propel::getConnection(ProductTableMap::DATABASE_NAME);
         $con->setAttribute(ConnectionWrapper::PROPEL_ATTR_CACHE_PREPARES, true);
+
         if ($this->isDebug()) {
+            // In debug mode, we have to initialize Tlog at this point, as this class uses Propel
+            Tlog::getInstance()->setLevel(Tlog::DEBUG);
+
             $serviceContainer->setLogger('defaultLogger', Tlog::getInstance());
             $con->useDebug(true);
         }