+++ core/lib/Thelia/Module/Validator/ModuleValidator.php
@@ -34,9 +34,6 @@ class ModuleValidator
 {
     protected $modulePath;
 
-    /** @var bool */
-    protected $isLoaded = false;
-
     /** @var ModuleDescriptorValidator */
     protected $moduleDescriptor;
 
@@ -51,9 +48,25 @@ class ModuleValidator
     /** @var array array of errors */
     protected $errors = [];
 
-    public function __construct($modulePath = null)
+    protected $moduleDirName;
+
+    /**
+     * @param string $modulePath the path of the module directory
+     * @param \Thelia\Core\Translation\Translator $translator FOR UNIT TEST PURPOSE ONLY
+     */
+    public function __construct($modulePath = null, $translator = null)
     {
+        $this->translator = $translator;
+
         $this->modulePath = $modulePath;
+
+        $this->moduleDirName = basename($this->modulePath);
+
+        $this->checkDirectoryStructure();
+
+        $this->loadModuleDescriptor();
+
+        $this->loadModuleDefinition();
     }
 
     /**
@@ -106,14 +119,6 @@ class ModuleValidator
     }
 
     /**
-     * @param \Thelia\Core\Translation\Translator $translator
-     */
-    public function setTranslator($translator)
-    {
-        $this->translator = $translator;
-    }
-
-    /**
      * @return \Thelia\Core\Translation\Translator
      */
     public function getTranslator()
@@ -123,23 +128,6 @@ class ModuleValidator
         }
 
         return $this->translator;
-    }
-
-
-    /**
-     * Loads the module description and configuration.
-     *
-     * @throws FileNotFoundException if modules does not exist or not contains module.xml or config.xml
-     */
-    public function load()
-    {
-        $this->checkDirectoryStructure();
-
-        $this->loadModuleDescriptor();
-
-        $this->loadModuleDefinition();
-
-        $this->isLoaded = true;
     }
 
     /**
@@ -153,14 +141,11 @@ class ModuleValidator
      */
     public function validate($checkCurrentVersion = true)
     {
-        if (false === $this->isLoaded) {
-            $this->load();
-        }
-
         if (null === $this->moduleDescriptor) {
             throw new \Exception(
                 $this->getTranslator()->trans(
-                    "The module definition has not been initialized."
+                    "The %name module definition has not been initialized.",
+                    [ '%name' => $this->moduleDirName ]
                 )
             );
         }
@@ -178,21 +163,30 @@ class ModuleValidator
     {
         if (false === file_exists($this->modulePath)) {
             throw new FileNotFoundException(
-                $this->getTranslator()->trans("Module directory doesn't exist.")
+                $this->getTranslator()->trans(
+                    "Module %name directory doesn't exists.",
+                    [ '%name' => $this->moduleDirName]
+                )
             );
         }
 
         $path = sprintf("%s/Config/module.xml", $this->modulePath);
         if (false === file_exists($path)) {
             throw new FileNotFoundException(
-                $this->getTranslator()->trans("Module should have a module.xml in the Config directory.")
+                $this->getTranslator()->trans(
+                    "Module %name should have a module.xml in the Config directory.",
+                    [ '%name' => $this->moduleDirName]
+                )
             );
         }
 
         $path = sprintf("%s/Config/config.xml", $this->modulePath);
         if (false === file_exists($path)) {
             throw new FileNotFoundException(
-                $this->getTranslator()->trans("Module should have a config.xml in the Config directory.")
+                $this->getTranslator()->trans(
+                    "Module %name should have a config.xml in the Config directory.",
+                    [ '%name' => $this->moduleDirName]
+                )
             );
         }
     }
@@ -217,15 +211,34 @@ class ModuleValidator
         if (null === $this->moduleDescriptor) {
             throw new \Exception(
                 $this->getTranslator()->trans(
-                    "The module descriptor has not been initialized."
+                    "The %name module descriptor has not been initialized.",
+                    [ '%name' => $this->moduleDirName ]
                 )
             );
         }
 
         $moduleDefinition = new ModuleDefinition();
 
-        $moduleDefinition->setCode(basename($this->modulePath));
-        $moduleDefinition->setNamespace((string)$this->moduleDescriptor->fullnamespace);
+        // Try to guess the proper module name, using the descriptor information.
+        $fullnamespace = trim((string)$this->moduleDescriptor->fullnamespace);
+
+        $namespaceComponents = explode("\\", $fullnamespace);
+
+        if (! isset($namespaceComponents[0]) || empty($namespaceComponents[0])) {
+            throw new ModuleException(
+                $this->getTranslator()->trans(
+                    "Unable to get module code from the fullnamespace element of the module descriptor: '%val'",
+                    [
+                        '%name' => $this->moduleDirName,
+                        '%val' => $fullnamespace
+                    ]
+                )
+            );
+        }
+
+        // Assume the module code is the first component of the declared namespace
+        $moduleDefinition->setCode($namespaceComponents[0]);
+        $moduleDefinition->setNamespace($fullnamespace);
         $moduleDefinition->setVersion((string)$this->moduleDescriptor->version);
 
         $this->getModuleLanguages($moduleDefinition);
@@ -253,8 +266,11 @@ class ModuleValidator
             if (!Version::test(Thelia::THELIA_VERSION, $this->moduleDefinition->getTheliaVersion(), false, ">=")) {
                 throw new ModuleException(
                     $this->getTranslator()->trans(
-                        "The module requires a version of Thelia >= %version",
-                        ['%version' => $this->moduleDefinition->getVersion()]
+                        "The module %name requires Thelia %version or newer",
+                        [
+                            '%name' => $this->moduleDirName,
+                            '%version' => $this->moduleDefinition->getVersion()
+                        ]
                     )
                 );
             }
@@ -270,7 +286,8 @@ class ModuleValidator
             if (version_compare($module->getVersion(), $this->moduleDefinition->getVersion(), '>=')) {
                 throw new ModuleException(
                     $this->getTranslator()->trans(
-                        "The module is already installed in the same or greater version."
+                        "The module %name is already installed in the same or greater version.",
+                        [ '%name' => $this->moduleDirName]
                     )
                 );
             }
@@ -312,8 +329,8 @@ class ModuleValidator
 
         if (count($errors) > 0) {
             $errorsMessage = $this->getTranslator()->trans(
-                'The module requires this activated modules : %modules',
-                ['%modules' => implode(', ', $errors)]
+                'To activate module %name, the following modules should be activated first: %modules',
+                ['%name' => $this->moduleDirName, '%modules' => implode(', ', $errors)]
             );
 
             throw new ModuleException($errorsMessage);
@@ -343,9 +360,9 @@ class ModuleValidator
 
         /** @var Module $module */
         foreach ($modules as $module) {
-            $validator = new ModuleValidator($module->getAbsoluteBaseDir());
             try {
-                $validator->load();
+                $validator = new ModuleValidator($module->getAbsoluteBaseDir());
+
                 $definition = $validator->getModuleDefinition();
                 $dependencies = $definition->getDependencies();