+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Autoload : the autoloader can be cached with Apc or XCache. See new index.php file.
 - Update : add missing API table creation
+- The default Tlog level is now TLog::ERROR instead of Tlog::DEBUG
 
 # 2.1.0-alpha2
 - Update Process :